Move Board class to separate file
